Л.Р 2.2 добавлена класс ArrayTabulatedFunction и реализован класс AbstractTabulatedFunction

+++ src/main/java/ru/ssau/tk/_KEPA_/_practice_/functions/TabulatedFunction.java
@@ -9,7 +9,7 @@ public interface TabulatedFunction extends MathFunction {
 
     void setY(int index, double value);
 
-    int indexOfX(double x);
+    double indexOfX(double x);
 
     int indexOfY(double y);
